feat: Start integrating inst.create permissions into repo/watch_branch

+++ src/aux-records/websockets/SplitInstRecordsStore.ts
@@ -8,6 +8,9 @@ import {
     InstRecordsStore,
     InstWithBranches,
     ReplaceUpdatesResult,
+    SaveBranchFailure,
+    SaveBranchResult,
+    SaveInstResult,
     StoredUpdates,
 } from './InstRecordsStore';
 import { TemporaryInstRecordsStore } from './TemporaryInstRecordsStore';
@@ -72,19 +75,29 @@ export class SplitInstRecordsStore implements InstRecordsStore {
         return info;
     }
 
-    async saveInst(inst: InstWithBranches): Promise<void> {
+    async saveInst(inst: InstWithBranches): Promise<SaveInstResult> {
         if (inst.recordName) {
-            await this._permanent.saveInst(inst);
+            const result = await this._permanent.saveInst(inst);
+            if (!result.success) {
+                return result;
+            }
             await this._temp.deleteAllInstBranchInfo(
                 inst.recordName,
                 inst.inst
             );
         }
+
+        return {
+            success: true,
+        };
     }
 
-    async saveBranch(branch: BranchRecord): Promise<void> {
+    async saveBranch(branch: BranchRecord): Promise<SaveBranchResult> {
         if (branch.recordName) {
-            await this._permanent.saveBranch(branch);
+            const result = await this._permanent.saveBranch(branch);
+            if (!result.success) {
+                return result;
+            }
             const info = await this._permanent.getBranchByName(
                 branch.recordName,
                 branch.inst,
@@ -102,6 +115,10 @@ export class SplitInstRecordsStore implements InstRecordsStore {
                 temporary: branch.temporary,
             });
         }
+
+        return {
+            success: true,
+        };
     }
 
     async getCurrentUpdates(
